Deployment/requirements.txt
streamlit 
pandas
seaborn
matplotlib
scikit-learn
Pillow